Adapting HPX plugin_factories to C++ modules
