+++ libs/full/parcelport_lci/include/hpx/parcelport_lci/sender_base.hpp
@@ -19,6 +19,7 @@
 #include <hpx/modules/functional.hpp>
 #include <hpx/modules/parcelset_base.hpp>
 #include <hpx/modules/plugin.hpp>
+#include <hpx/modules/plugin_factories.hpp>
 #include <hpx/modules/resource_partitioner.hpp>
 #include <hpx/modules/runtime_configuration.hpp>
 #include <hpx/modules/runtime_local.hpp>
@@ -28,7 +29,6 @@
 #include <hpx/parcelset/parcelport_connection.hpp>
 #include <hpx/parcelset/parcelport_impl.hpp>
 #include <hpx/parcelset/parcelset_fwd.hpp>
-#include <hpx/plugin_factories/parcelport_factory.hpp>
 
 #include <hpx/modules/parcelset_base.hpp>
 #include <hpx/parcelset/parcelport_connection.hpp>
