+++ libs/full/parcelport_mpi/CMakeLists.txt
@@ -38,7 +38,7 @@ add_hpx_module(
   COMPAT_HEADERS ${parcelport_mpi_compat_headers}
   DEPENDENCIES hpx_core hpx_dependencies_boost Mpi::mpi
   MODULE_DEPENDENCIES hpx_actions hpx_command_line_handling hpx_parcelset
-                      hpx_parcelset_base
+                      hpx_parcelset_base hpx_plugin_factories
   CMAKE_SUBDIRS examples tests
 )
