+++ libs/full/parcelports/cmake/templates/static_parcelports.hpp.in
@@ -12,7 +12,7 @@
 #include <hpx/config.hpp>
 
 #if defined(HPX_HAVE_NETWORKING)
-#include <hpx/plugin_factories/parcelport_factory_base.hpp>
+#include <hpx/modules/plugin_factories.hpp>
 
 #include <vector>
