+++ libs/full/parcelset/include/hpx/parcelset/init_parcelports.hpp
@@ -9,7 +9,7 @@
 #include <hpx/config.hpp>
 
 #if defined(HPX_HAVE_NETWORKING)
-#include <hpx/plugin_factories/parcelport_factory_base.hpp>
+#include <hpx/modules/plugin_factories.hpp>
 
 #include <vector>
