+++ libs/full/parcelset/include/hpx/parcelset/parcelhandler.hpp
@@ -23,8 +23,8 @@
 #include <hpx/modules/components_base.hpp>
 #include <hpx/modules/naming_base.hpp>
 #include <hpx/modules/parcelset_base.hpp>
+#include <hpx/modules/plugin_factories.hpp>
 #include <hpx/parcelset/parcelset_fwd.hpp>
-#include <hpx/plugin_factories/parcelport_factory_base.hpp>
 
 #include <algorithm>
 #include <atomic>
