+++ libs/full/plugin_factories/include/hpx/plugin_factories/binary_filter_factory.hpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2007-2024 Hartmut Kaiser
+//  Copyright (c) 2007-2026 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -8,9 +8,8 @@
 
 #include <hpx/config.hpp>
 #include <hpx/modules/ini.hpp>
-#include <hpx/modules/preprocessor.hpp>
-
 #include <hpx/plugin_factories/binary_filter_factory_base.hpp>
+#include <hpx/plugin_factories/macros.hpp>
 #include <hpx/plugin_factories/plugin_registry.hpp>
 #include <hpx/plugin_factories/unique_plugin_name.hpp>
 
@@ -25,7 +24,7 @@ namespace hpx::plugins {
     ///
     /// \tparam BinaryFilter The message handler type this factory should be
     ///                        responsible for.
-    template <typename BinaryFilter>
+    HPX_CXX_EXPORT template <typename BinaryFilter>
     struct binary_filter_factory : binary_filter_factory_base
     {
         /// \brief Construct a new factory instance
@@ -76,15 +75,3 @@ namespace hpx::plugins {
         bool isenabled_;
     };
 }    // namespace hpx::plugins
-
-///////////////////////////////////////////////////////////////////////////////
-/// This macro is used create and to register a minimal component factory with
-/// Hpx.Plugin.
-#define HPX_REGISTER_BINARY_FILTER_FACTORY(BinaryFilter, pluginname)           \
-    HPX_REGISTER_BINARY_FILTER_FACTORY_BASE(                                   \
-        hpx::plugins::binary_filter_factory<BinaryFilter>, pluginname)         \
-    HPX_DEF_UNIQUE_PLUGIN_NAME(                                                \
-        hpx::plugins::binary_filter_factory<BinaryFilter>, pluginname)         \
-    template struct hpx::plugins::binary_filter_factory<BinaryFilter>;         \
-    HPX_REGISTER_PLUGIN_REGISTRY_2(BinaryFilter, pluginname)                   \
-    /**/
