+++ libs/full/plugin_factories/include/hpx/plugin_factories/binary_filter_factory_base.hpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2007-2021 Hartmut Kaiser
+//  Copyright (c) 2007-#include <hpx/plugin_factories/macros.hpp>2026 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -9,7 +9,7 @@
 #include <hpx/config.hpp>
 #include <hpx/modules/plugin.hpp>
 #include <hpx/modules/serialization.hpp>
-
+#include <hpx/plugin_factories/macros.hpp>
 #include <hpx/plugin_factories/plugin_factory_base.hpp>
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -18,7 +18,8 @@ namespace hpx::plugins {
     ///////////////////////////////////////////////////////////////////////////
     /// The \a plugin_factory_base has to be used as a base class for all
     /// plugin factories.
-    struct HPX_EXPORT binary_filter_factory_base : plugin_factory_base
+    HPX_CXX_EXPORT struct HPX_EXPORT binary_filter_factory_base
+      : plugin_factory_base
     {
         ~binary_filter_factory_base() override = default;
 
@@ -30,14 +31,3 @@ namespace hpx::plugins {
             serialization::binary_filter* next_filter = nullptr) = 0;
     };
 }    // namespace hpx::plugins
-
-///////////////////////////////////////////////////////////////////////////////
-/// This macro is used to register the given binary filter factory with
-/// Hpx.Plugin. This macro has to be used for each of the binary filter
-/// factories.
-#define HPX_REGISTER_BINARY_FILTER_FACTORY_BASE(FactoryType, pluginname)       \
-    HPX_PLUGIN_EXPORT(HPX_PLUGIN_PLUGIN_PREFIX,                                \
-        hpx::plugins::plugin_factory_base, FactoryType, pluginname, factory)   \
-    HPX_INIT_REGISTRY_PLUGIN_FACTORY_STATIC(                                   \
-        HPX_PLUGIN_PLUGIN_PREFIX, pluginname, factory)                         \
-    /**/
